+++ processor/src/test/fixtures/bad_input/com/example/inject/Jsr330ScopedNonDaggerModel.java
@@ -3,12 +3,11 @@ package com.example.inject;
 import arez.annotations.Action;
 import arez.annotations.ArezComponent;
 import arez.annotations.Feature;
-import arez.annotations.InjectMode;
 import javax.inject.Singleton;
 
 @Singleton
-@ArezComponent( inject = InjectMode.NONE, dagger = Feature.DISABLE )
-public abstract class ScopePresentInjectDisabledModel
+@ArezComponent( dagger = Feature.DISABLE )
+public abstract class Jsr330ScopedNonDaggerModel
 {
   @Action
   void myAction()